Created settings.json file, you can create specific settings for
each repository on the file `settings.json`, which should be put on
the aimed repository root directory of repository.

The `tags` is a list of remote tags and the version of Sublime they
are aimed to be used. For example, `tags = 3143, 3147` indicates
the repository has the following git tags called `3143` and `3147`.